Search Physician Office Visit Costs Data sets by Specialty

Returns information pertaining to 83 datasets separated by medical specialty, describing physician office visit costs per zip code for new and established patients.

Usage

search_datasets(specialty = NULL, keyword = NULL)

Source

Physician Office Visit Costs (Data.CMS.gov)

Arguments

specialty

search for exact medical specialty, e.g. 'cardiology'

keyword

search for partial matches, e.g. 'medicine'

Value

A tidytable containing the specialty of the provider, the title of the dataset, the dates that the dataset was issued, modified, and released, and the link to download the csv file

Medical Specialties Available

Click to View

search_datasets()$specialty
#>  [1] "addiction medicine"                                     
#>  [2] "advanced heart failure and transplant cardiology"       
#>  [3] "allergy/ immunology"                                    
#>  [4] "anesthesiology"                                         
#>  [5] "cardiac surgery"                                        
#>  [6] "cardiology"                                             
#>  [7] "certified clinical nurse specialist"                    
#>  [8] "certified nurse midwife"                                
#>  [9] "certified registered nurse anesthetist (crna)"          
#> [10] "clinic or group practice"                               
#> [11] "clinical cardiac electrophysiology"                     
#> [12] "clinical laboratory"                                    
#> [13] "colorectal surgery (proctology)"                        
#> [14] "critical care (intensivists)"                           
#> [15] "dentist"                                                
#> [16] "dermatology"                                            
#> [17] "diagnostic radiology"                                   
#> [18] "emergency medicine"                                     
#> [19] "endocrinology"                                          
#> [20] "family practice"                                        
#> [21] "gastroenterology"                                       
#> [22] "general practice"                                       
#> [23] "general surgery"                                        
#> [24] "geriatric medicine"                                     
#> [25] "geriatric psychiatry"                                   
#> [26] "gynecological oncology"                                 
#> [27] "hand surgery"                                           
#> [28] "hematology"                                             
#> [29] "hematology-oncology"                                    
#> [30] "hematopoietic cell transplantation and cellular therapy"
#> [31] "hospice and palliative care"                            
#> [32] "hospitalist"                                            
#> [33] "infectious disease"                                     
#> [34] "internal medicine"                                      
#> [35] "interventional cardiology"                              
#> [36] "interventional pain management"                         
#> [37] "interventional radiology"                               
#> [38] "licensed clinical social worker"                        
#> [39] "mammography center"                                     
#> [40] "maxillofacial surgery"                                  
#> [41] "medical genetics and genomics"                          
#> [42] "medical oncology"                                       
#> [43] "medical toxicology"                                     
#> [44] "nephrology"                                             
#> [45] "neurology"                                              
#> [46] "neuropsychiatry"                                        
#> [47] "neurosurgery"                                           
#> [48] "nuclear medicine"                                       
#> [49] "nurse practitioner"                                     
#> [50] "obstetrics &amp; gynecology"                            
#> [51] "ophthalmology"                                          
#> [52] "optometry"                                              
#> [53] "oral surgery (dentist only)"                            
#> [54] "orthopedic surgery"                                     
#> [55] "osteopathic manipulative medicine"                      
#> [56] "otolaryngology"                                         
#> [57] "pain management"                                        
#> [58] "pathology"                                              
#> [59] "pediatric medicine"                                     
#> [60] "peripheral vascular disease"                            
#> [61] "physical medicine and rehabilitation"                   
#> [62] "physical therapist in private practice"                 
#> [63] "physician assistant"                                    
#> [64] "plastic and reconstructive surgery"                     
#> [65] "podiatry"                                               
#> [66] "preventive medicine"                                    
#> [67] "psychiatry"                                             
#> [68] "psychologist, clinical"                                 
#> [69] "public health or welfare agency"                        
#> [70] "pulmonary disease"                                      
#> [71] "radiation oncology"                                     
#> [72] "registered dietitian or nutrition professional"         
#> [73] "rheumatology"                                           
#> [74] "sleep medicine"                                         
#> [75] "speech language pathologist"                            
#> [76] "sports medicine"                                        
#> [77] "surgical oncology"                                      
#> [78] "thoracic surgery"                                       
#> [79] "undefined physician type"                               
#> [80] "undersea and hyperbaric medicine"                       
#> [81] "unknown supplier/provider specialty"                    
#> [82] "urology"                                                
#> [83] "vascular surgery"

Examples

# The `specialty` argument will only return _exact_ matches
search_datasets(specialty = "cardiac surgery")
#> # A tidytable: 1 × 6
#>   specialty       title                 issued     modified   released   csv_url
#>   <chr>           <chr>                 <date>     <date>     <date>     <chr>  
#> 1 cardiac surgery Cardiac Surgery Offi… 2022-07-11 2022-07-11 2022-07-14 https:…

# The `keyword` argument will return _partial_ matches
search_datasets(keyword = "medicine")
#> # A tidytable: 12 × 6
#>    specialty                      title issued     modified   released   csv_url
#>    <chr>                          <chr> <date>     <date>     <date>     <chr>  
#>  1 addiction medicine             Addi… 2022-07-11 2022-07-11 2022-07-14 https:…
#>  2 emergency medicine             Emer… 2022-07-11 2022-07-11 2022-07-14 https:…
#>  3 geriatric medicine             Geri… 2022-07-11 2022-07-11 2022-07-14 https:…
#>  4 internal medicine              Inte… 2022-07-11 2022-07-11 2022-07-14 https:…
#>  5 nuclear medicine               Nucl… 2022-07-11 2022-07-11 2022-07-14 https:…
#>  6 osteopathic manipulative medi… Oste… 2022-07-11 2022-07-11 2022-07-14 https:…
#>  7 pediatric medicine             Pedi… 2022-07-11 2022-07-11 2022-07-14 https:…
#>  8 physical medicine and rehabil… Phys… 2022-07-11 2022-07-11 2022-07-14 https:…
#>  9 preventive medicine            Prev… 2022-07-11 2022-07-11 2022-07-14 https:…
#> 10 sleep medicine                 Slee… 2022-07-11 2022-07-11 2022-07-14 https:…
#> 11 sports medicine                Spor… 2022-07-11 2022-07-11 2022-07-14 https:…
#> 12 undersea and hyperbaric medic… Unde… 2022-07-11 2022-07-11 2022-07-14 https:…

# Calling the function with no arguments will return the entire dataset
search_datasets()
#> # A tidytable: 83 × 6
#>    specialty                      title issued     modified   released   csv_url
#>    <chr>                          <chr> <date>     <date>     <date>     <chr>  
#>  1 addiction medicine             Addi… 2022-07-11 2022-07-11 2022-07-14 https:…
#>  2 advanced heart failure and tr… Adva… 2022-07-11 2022-07-11 2022-07-14 https:…
#>  3 allergy/ immunology            Alle… 2022-07-11 2022-07-11 2022-07-14 https:…
#>  4 anesthesiology                 Anes… 2022-07-11 2022-07-11 2022-07-14 https:…
#>  5 cardiac surgery                Card… 2022-07-11 2022-07-11 2022-07-14 https:…
#>  6 cardiology                     Card… 2022-07-11 2022-07-11 2022-07-14 https:…
#>  7 certified clinical nurse spec… Cert… 2022-07-11 2022-07-11 2022-07-14 https:…
#>  8 certified nurse midwife        Cert… 2022-07-11 2022-07-11 2022-07-14 https:…
#>  9 certified registered nurse an… Cert… 2022-07-11 2022-07-11 2022-07-14 https:…
#> 10 clinic or group practice       Clin… 2022-07-11 2022-07-11 2022-07-14 https:…
#> # ℹ 73 more rows